feat: stabilize RLM-DSL v0.2 and implement spatial primitives
- Introduce TASK-based top-level structure in AST and parser.
- Implement deterministic spatial primitives: AFTER_TEXT, AFTER_REGEX, MATCH_GROUP, CAPTURE_REGEX_GROUP, VALUE_AFTER_DELIM.
- Add STRICT linter rule LINT_OFFSET_ARITHMETIC_FORBIDDEN to prevent unreliable character counting.
- Implement Anka-style AutoRepair engine with rewrite rules for offset arithmetic and undefined identifiers.
- Update benchmark harness with detailed summary reporting and 100% pass rate target logic.
- Expand documentation (SPEC.md, syntax_guide.md) to reflect v0.2 canonical form.
- Achieve 90%+ pass rate on Security and Combined Workflow suites.

+++ internal/ops/pure/pure_test.go
@@ -56,7 +56,7 @@ func TestPureOps(t *testing.T) {
 	vj := runtime.Value{Kind: runtime.KindText, V: hj}
 	pj, _ := JSONParse(s, vj)
 	gj, _ := JSONGet(s, pj, "a.b")
-	if gj.V.(float64) != 1 {
+	if gj.V.(int) != 1 {
 		t.Errorf("expected 1, got %v", gj.V)
 	}
